Reduce transfer overhead.
For export, we had been using shutil.copy2 for copying files.  Prior to
Python 3.8, this internally uses a plain Python loop.  Python 3.8 and
later use posix.sendfile.  In a linux docker under Windows, shutil.copy2
is slow; the sendfile method is worse than before by a factor of 3 or so
in my test setup.  By using a subprocess and cp, in this environment we
get roughly a factor of 2 faster than the pre 3.8 shutil function.

Further, python has a lot of overhead (in a linux docker under Windows)
for each read call.  By aggregating reads together (see tifftools), we
can save some of this overhead.