Check settings for valid regex
The new settings to hide metadata/disable redaction are configured as
lists of regular expressions. Metadata keys that match these patterns
will either be hidden or unable to be redacted. This commit catches
invalid regular expressions in those settings, ensuring the metadata secion
of the UI loads even if there are errors in configuration. For every invalid
regular expression in the settings, a message is printed to the console with
additional information about the error.