Transition to dictionary-based config
Configuration that drives enabling the redaction control and
hiding metadata now uses a dictionary structure. The keys for
these dictionaries are regular expressions representing metadata
keys, and the values in the dictionary and regular expressions
representing metadata values. If a metadata key is flagged to be
hidden or have redaction disabled, AND the value of that metadata field
matches an expected value (usually numeric), then its ok to disable the
control for or hide that metadata. This means that any malformed metadata
will still be available to end users for redaction.