Add functionality to edit metadata values
Includes event handling for changes to the recently added textbox, which
shows up if a user has a redaction reason selected for a piece of metadata.
Now users can use this textbox to make changes persist to the item When
redaction is performed on an item, metadata values are actually edited on
the image.