Pin EasyOcr to 1.4
Also remove tesseract from the Dockerfile/setup.py in favor of EasyOCR. Pinning
EasyOCR to this specific version in order to avoid vulnerability
CVE-2021-34552. EasyOCR 1.4.1 is dependent on Pillow<8.3.0, while 1.4.0
is not. Since we need Pillow>=8.3.0 to avoid the vulnerability, we'll use
EasyOCR 1.4.0 for now. There is a fix in the source code for EasyOCR as of
9/22/2021. Once that is released we can remove the specified version.