Matching API
Add settings for API URL matching.  Add a testing endpoint to simulate
results.  Add a matching endpoint to call the API.  When a file is
ingested and unfiled, apply the matching queries to file it.

For both unfiled files and available to process files, offer to look up
values from the API and refile the file.  If a database API is
configured, there is a new button "Lookup" which opens a dialog showing
the label image (if any) and having text entry boxes for the lookup
fields.  Clicking "Lookup" will print matches.  Selecting "Apply" next
to a match will refile the image using that match.