Add a system endpoint to report the complete config and settings
This is handy for debugging, but might be considered too insecure
(though all of the information can be obtained in other ways).

It would have been nice to track the files actually used for girder
config, but those are loaded before plugins, so it is not actually
possible without _reloading_ the config (not done here).

Closes #352.